Move User data class under java/data/user folder
