+++ android/app/src/main/java/com/bounswe2020group3/paperlayer/data/user/User.kt
@@ -1,0 +1,15 @@
+package com.bounswe2020group3.paperlayer.data.user
+
+import com.bounswe2020group3.paperlayer.profile.data.Profile
+import com.squareup.moshi.Json
+
+data class User (
+    @field:Json(name = "id")
+    var id: Int,
+    @field:Json(name = "username")
+    var username: String,
+    @field:Json(name = "email")
+    var email: String,
+    @field:Json(name = "profile")
+    var profile: List<Profile>,
+)
