+++ android/app/src/main/java/com/bounswe2020group3/paperlayer/profile/ProfileModel.kt
@@ -3,7 +3,7 @@ package com.bounswe2020group3.paperlayer.profile
 import com.bounswe2020group3.paperlayer.profile.data.data.AuthToken
 import com.bounswe2020group3.paperlayer.network.RetrofitProvider
 import com.bounswe2020group3.paperlayer.profile.data.Profile
-import com.bounswe2020group3.paperlayer.profile.data.User
+import com.bounswe2020group3.paperlayer.data.user.User
 import com.bounswe2020group3.paperlayer.util.Session
 import io.reactivex.Observable
 import io.reactivex.Single
