+++ android/app/src/main/java/com/bounswe2020group3/paperlayer/profile/edit/ProfileEditContract.kt
@@ -2,7 +2,7 @@ package com.bounswe2020group3.paperlayer.profile.edit
 
 import com.bounswe2020group3.paperlayer.mvp.Mvp
 import com.bounswe2020group3.paperlayer.profile.data.Profile
-import com.bounswe2020group3.paperlayer.profile.data.User
+import com.bounswe2020group3.paperlayer.data.user.User
 
 interface ProfileEditContract {
     interface Presenter : Mvp.Presenter<View> {
