+++ android/app/src/main/java/com/bounswe2020group3/paperlayer/profile/list/UserListContract.kt
@@ -1,7 +1,7 @@
 package com.bounswe2020group3.paperlayer.profile.list
 
 import com.bounswe2020group3.paperlayer.mvp.Mvp
-import com.bounswe2020group3.paperlayer.profile.data.User
+import com.bounswe2020group3.paperlayer.data.user.User
 
 interface UserListContract {
     interface Presenter: Mvp.Presenter<View> {
