+++ android/app/src/main/java/com/bounswe2020group3/paperlayer/profile/user/UserContract.kt
@@ -1,8 +1,7 @@
 package com.bounswe2020group3.paperlayer.profile.user
 
 import com.bounswe2020group3.paperlayer.mvp.Mvp
-import com.bounswe2020group3.paperlayer.profile.data.Profile
-import com.bounswe2020group3.paperlayer.profile.data.User
+import com.bounswe2020group3.paperlayer.data.user.User
 
 interface UserContract {
     interface Presenter : Mvp.Presenter<View> {
