+++ android/app/src/main/java/com/bounswe2020group3/paperlayer/profile/user/UserFragment.kt
@@ -10,7 +10,7 @@ import android.widget.Toast
 import androidx.navigation.Navigation
 import com.bounswe2020group3.paperlayer.MainActivity
 import com.bounswe2020group3.paperlayer.R
-import com.bounswe2020group3.paperlayer.profile.data.User
+import com.bounswe2020group3.paperlayer.data.user.User
 import com.squareup.picasso.Picasso
 import kotlinx.android.synthetic.main.fragment_user.*
 import javax.inject.Inject
